1. 增加删除学校，班级等校验 2. 删除学生时，级联删除成绩，关系 3. 首页相关接口增加缓存。 -陈礼松

+++ nuo-api/src/main/java/com/yinuo/service/ClassesService.java
@@ -1,19 +1,21 @@
 package com.yinuo.service;
 
-import java.util.Date;
-import java.util.List;
-
+import com.alibaba.fastjson.JSON;
+import com.yinuo.bean.Classes;
 import com.yinuo.bean.Constant;
+import com.yinuo.bean.User;
+import com.yinuo.exception.InvalidArgumentException;
+import com.yinuo.mapper.ClassesMapper;
+import com.yinuo.util.CommonUtil;
+import com.yinuo.util.DateTool;
+import com.yinuo.util.JedisUtil;
+import com.yinuo.util.MergerUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.yinuo.bean.Classes;
-import com.yinuo.bean.User;
-import com.yinuo.mapper.ClassesMapper;
-import com.yinuo.util.CommonUtil;
-import com.yinuo.util.DateTool;
-import com.yinuo.util.MergerUtil;
+import java.util.Date;
+import java.util.List;
 
 @Transactional(rollbackFor = Exception.class)
 @Service
@@ -21,6 +23,12 @@ public class ClassesService {
 	
 	@Autowired
 	private ClassesMapper classesMapper;
+
+	@Autowired
+	private StudentService studentService;
+
+	@Autowired
+	private JedisUtil jedis;
 	
 	public long insert(User loginUser, Classes classes) {
 		loginUser.checkLevel(Constant.Role.School, classes.getSchoolId(), 0L);
@@ -34,7 +42,15 @@ public class ClassesService {
 	}
 	
 	public Classes selectOne(long id) {
-		return classesMapper.selectOne(id);
+		Classes classes = jedis.get(Constant.JedisNames.ClassId + id, Classes.class);
+		if(classes == null) {
+			classes = classesMapper.selectOne(id);
+			if(classes != null) {
+				jedis.set(Constant.JedisNames.ClassId + id, JSON.toJSONString(classes), Constant.JedisNames.OneDay);
+			}
+		}
+
+		return classes;
 	}
 	
 	public void delete (User loginUser, long id) {
@@ -42,7 +58,12 @@ public class ClassesService {
 		CommonUtil.checkNull(classes, "找不到该班级");
 		loginUser.checkLevel(Constant.Role.School, classes.getSchoolId(), classes.getId());
 
+		if(studentService.countByClassid(id) > 0) {
+			throw new InvalidArgumentException("请先删除学生");
+		}
+
 		classesMapper.delete(id);
+		jedis.del(Constant.JedisNames.ClassId + id);
 	}
 	
 	//只能自己修改自己的数据场景
@@ -53,6 +74,7 @@ public class ClassesService {
 
 		loginUser.checkLevel(Constant.Role.School, classes.getSchoolId(), 0L);
 		classesMapper.update(classes);
+		jedis.del(Constant.JedisNames.ClassId + classes.getId());
 	}
 	
 	public List<Classes> selectListByIds(List<Long> ids) {
