+++ nuo-api/src/main/java/com/yinuo/service/SchoolService.java
@@ -1,11 +1,15 @@
 package com.yinuo.service;
 
+import com.alibaba.fastjson.JSON;
+import com.yinuo.bean.Classes;
 import com.yinuo.bean.School;
 import com.yinuo.bean.Constant;
 import com.yinuo.bean.User;
+import com.yinuo.exception.InvalidArgumentException;
 import com.yinuo.mapper.SchoolMapper;
 import com.yinuo.util.CommonUtil;
 import com.yinuo.util.DateTool;
+import com.yinuo.util.JedisUtil;
 import com.yinuo.util.MergerUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,6 +24,12 @@ public class SchoolService {
 	
 	@Autowired
 	private SchoolMapper schoolMapper;
+
+	@Autowired
+	private ClassesService classesService;
+
+	@Autowired
+	private JedisUtil jedis;
 	
 	public long insert(User loginUser, School school) {
 		loginUser.checkLevel(Constant.Role.Manager, 0L, 0L);
@@ -33,13 +43,25 @@ public class SchoolService {
 	}
 	
 	public School selectOne(long id) {
-		return schoolMapper.selectOne(id);
+
+		School school = jedis.get(Constant.JedisNames.SchoolId + id, School.class);
+		if(school == null) {
+			school = schoolMapper.selectOne(id);
+			if(school != null) {
+				jedis.set(Constant.JedisNames.SchoolId + id, JSON.toJSONString(school), Constant.JedisNames.OneDay);
+			}
+		}
+		return school;
 	}
 	
 	public void delete (User loginUser, long id) {
 		loginUser.checkLevel(Constant.Role.Manager, 0L, 0L);
 
+		if(classesService.countBySchoolId(id, 0) > 0) {
+			throw new InvalidArgumentException("请先删除班级");
+		}
 		schoolMapper.delete(id);
+		jedis.del(Constant.JedisNames.SchoolId + id);
 	}
 	
 	//只能自己修改自己的数据场景
@@ -50,6 +72,7 @@ public class SchoolService {
 
 		loginUser.checkLevel(Constant.Role.Manager, 0L, 0L);
 		schoolMapper.update(school);
+		jedis.del(Constant.JedisNames.SchoolId + school.getId());
 	}
 	
 	public List<School> selectListByIds(List<Long> ids) {
